New colour scheme for tempertatures and tools
The current temperature backgrounds now reflect the difference
between the current temp and the set point.  If the temp is
within 1% of the set point, the background is green.  If >1% but
less than 3%, the background is yellow.  If >3% the background is
red.  The set point is determined by whether the tool is in active
or standby state.  The backgrounds of the active and standby temps
are also set based on their state.  Whichever one is active, is set
to green and the other is set to dark grey.

Changes:

 * Added new entries to ColourSchemes.
 * Added an enum to UserInterface that mirrors M408 hstat.
 * Added code to UpdateCurrentTemperature to set background
   colours.
 * Changed prototype of UpdateHeaterStatus to use the new enum.
 * Added code to UpdateHeaterStatus to set active and standby
   background colours.